+++ app/controllers/comments_controller.rb
@@ -3,73 +3,71 @@
 #   the COPYRIGHT file.
 
 class CommentsController < ApplicationController
-  include ApplicationHelper
-  before_action :authenticate_user!, :except => [:index]
+  before_action :authenticate_user!, except: :index
 
-  respond_to :html,
-             :mobile,
-             :json
+  respond_to :html, :mobile, :json
 
   rescue_from ActiveRecord::RecordNotFound do
-    render :nothing => true, :status => 404
+    render nothing: true, status: 404
   end
 
   def create
-    post = current_user.find_visible_shareable_by_id(Post, params[:post_id])
-    @comment = current_user.comment!(post, params[:text]) if post
-
+    @comment = CommentService.new(post_id: params[:post_id], text: params[:text], user: current_user).create_comment
     if @comment
-      respond_to do |format|
-        format.json{ render :json => CommentPresenter.new(@comment), :status => 201 }
-        format.html{ render :nothing => true, :status => 201 }
-        format.mobile{ render :partial => 'comment', :locals => {:post => @comment.post, :comment => @comment} }
-      end
+      respond_create_success
     else
-      render :nothing => true, :status => 422
+      render nothing: true, status: 404
     end
   end
 
   def destroy
-    @comment = Comment.find(params[:id])
-    if current_user.owns?(@comment) || current_user.owns?(@comment.parent)
-      current_user.retract(@comment)
-      respond_to do |format|
-        format.js { render :nothing => true, :status => 204 }
-        format.json { render :nothing => true, :status => 204 }
-        format.mobile{ redirect_to :back }
-      end
+    service = CommentService.new(comment_id: params[:id], user: current_user)
+    if service.destroy_comment
+      respond_destroy_success
     else
-      respond_to do |format|
-        format.mobile { redirect_to :back }
-        format.any(:js, :json) {render :nothing => true, :status => 403}
-      end
+      respond_destroy_error
     end
   end
 
   def new
     respond_to do |format|
-      format.mobile { render :layout => false }
+      format.mobile { render layout: false }
     end
   end
 
   def index
-    find_post
-    raise(ActiveRecord::RecordNotFound.new) unless @post
-
-    @comments = @post.comments.for_a_stream
+    service = CommentService.new(post_id: params[:post_id], user: current_user)
+    @post = service.post
+    @comments = service.comments
     respond_with do |format|
-      format.json  { render :json => CommentPresenter.as_collection(@comments), :status => 200 }
-      format.mobile{render :layout => false}
+      format.json  { render json: CommentPresenter.as_collection(@comments), status: 200 }
+      format.mobile { render layout: false }
     end
   end
 
   private
 
-  def find_post
-    if user_signed_in?
-      @post = current_user.find_visible_shareable_by_id(Post, params[:post_id])
-    else
-      @post = Post.find_by_id_and_public(params[:post_id], true)
+  def respond_create_success
+    respond_to do |format|
+      format.json { render json: CommentPresenter.new(@comment), status: 201 }
+      format.html { render nothing: true, status: 201 }
+      format.mobile { render partial: "comment", locals: {post: @comment.post, comment: @comment} }
+    end
+  end
+
+  def respond_destroy_success
+    respond_to do |format|
+      format.mobile { redirect_to :back }
+      format.js { render nothing: true, status: 204 }
+      format.json { render nothing: true, status: 204 }
+    end
+  end
+
+  def respond_destroy_error
+    respond_to do |format|
+      format.mobile { redirect_to :back }
+      format.js { render nothing: true, status: 403 }
+      format.json { render nothing: true, status: 403 }
     end
   end
 end